Reconcile 2023 changes
