+++ frontend/.eslintrc.json
@@ -22,6 +22,8 @@
     "no-param-reassign": 0,
     "no-console": 0,
     "react/prop-types": "off",
+    "react/jsx-uses-react": "off",
+    "react/react-in-jsx-scope": "off",
     "react/jsx-props-no-spreading": "off",
     "jsx-a11y/label-has-associated-control": "off",
     "react/jsx-no-bind": "off",
