+++ frontend/src/components/ContextMenu.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import Menu from "@material-ui/core/Menu";
 
 /**
