+++ frontend/src/components/Thumbnail.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import Skeleton from "react-loading-skeleton";
 
 /**
