+++ frontend/src/components/Triangle.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 export default function Triangle({
   triangleWidth,
   triangleHeight,
