+++ frontend/src/context/PlayScenarioContext.jsx
