+++ frontend/src/context/PlayScenarioContextProvider.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { useGet } from "../hooks/crudHooks";
 import PlayScenarioContext from "./PlayScenarioContext";
