+++ frontend/src/context/ScenarioContext.jsx
