+++ frontend/src/context/ToolbarContextProvider.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import ToolbarContext from "./ToolbarContext";
 
 /**
