+++ package.json
@@ -13,7 +13,7 @@
     "contributors:generate": "all-contributors generate"
   },
   "dependencies": {
-    "concurrently": "^8.2.1",
+    "concurrently": "^8.2.2",
     "npm": "6.14",
     "papaparse": "^5.4.1",
     "react": "^18.1.0",
